+++ src/renderer/stores/connections.ts
@@ -163,7 +163,8 @@ export const useConnectionsStore = defineStore('connections', {
                   uid: connection.uid,
                   client: connection.client,
                   icon: conn.icon,
-                  name: conn.name
+                  name: conn.name,
+                  hasCustomIcon: conn.hasCustomIcon
                };
             }
             return conn;